requirements.txt
bcrypt==4.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
more-itertools==10.7.0
numpy==2.3.0
pillow==11.3.0
pygame==2.6.1
PyOpenGL==3.1.9
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
six==1.17.0
